refactor: introduce network domain for related files
