+++ src/__swaps__/screens/Swap/components/AnimatedChainImage.android.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 import { Image, View } from 'react-native';
 
-import { getChainBadgeStyles } from '@/components/coin-icon/ChainImage';
 import { useColorMode } from '@/design-system';
+import { getChainBadgeStyles } from '@/features/network/components/ChainImage';
 import { useBackendNetworksStore } from '@/state/backendNetworks/backendNetworks';
 import { ChainId } from '@/state/backendNetworks/types';
 import { useSwapsStore } from '@/state/swaps/swapsStore';
