+++ src/__swaps__/screens/Swap/components/TokenList/ChainSelection.tsx
@@ -9,8 +9,8 @@ import { type UserAssetFilter } from '@/__swaps__/types/assets';
 import { type TokenToBuyListItem } from '@/__swaps__/types/search';
 import { analytics } from '@/analytics';
 import { GestureHandlerButton } from '@/components/buttons/GestureHandlerButton';
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { AnimatedText, Bleed, Box, globalColors, Inline, Text, TextIcon, useColorMode } from '@/design-system';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { useSharedValueState } from '@/hooks/reanimated/useSharedValueState';
 import { useAccountAccentColor } from '@/hooks/useAccountAccentColor';
