+++ src/components/L2Disclaimer.js
@@ -3,8 +3,8 @@ import { Platform } from 'react-native';
 
 import RadialGradient from 'react-native-radial-gradient';
 
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import Divider from '@/components/Divider';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { opacity } from '@/framework/ui/utils/opacity';
 import * as i18n from '@/languages';
 import { useBackendNetworksStore } from '@/state/backendNetworks/backendNetworks';
