+++ src/components/coin-icon/TwoCoinsIcon.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { Box } from '@/design-system';
 import type { ParsedAddressAsset } from '@/entities/tokens';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { type ChainId } from '@/state/backendNetworks/types';
 import { useTheme } from '@/theme/ThemeContext';
