+++ src/components/context-menu-buttons/ChainContextMenu.tsx
@@ -1,8 +1,8 @@
 import React, { useCallback, useMemo } from 'react';
 
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { ContextMenuButton } from '@/components/context-menu';
 import { Bleed, Box, Inline, Text, type TextProps } from '@/design-system';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { showActionSheetWithOptions } from '@/framework/ui/utils/actionsheet';
 import * as i18n from '@/languages';
 import { useUserAssetsStore } from '@/state/assets/userAssets';
