+++ src/components/rainbow-toast/icons/SendToastIcon.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { StyleSheet, View } from 'react-native';
 
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { TOAST_ICON_SIZE } from '@/components/rainbow-toast/constants';
 import { ToastSFSymbolIcon } from '@/components/rainbow-toast/ToastSFSymbolIcon';
 import type { RainbowToast } from '@/components/rainbow-toast/types';
 import { RainbowImage } from '@/components/RainbowImage';
 import { TransactionStatus } from '@/entities/transactions';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { ChainId } from '@/state/backendNetworks/types';
 
 export const SendToastIcon = ({ toast, size = TOAST_ICON_SIZE }: { toast: RainbowToast; size?: number }) => {
