+++ src/features/network/components/SearchBar.tsx
@@ -9,7 +9,8 @@ import { Box, globalColors, Text, useColorMode } from '@/design-system';
 import { useForegroundColor } from '@/design-system/color/useForegroundColor';
 import { opacity } from '@/framework/ui/utils/opacity';
 import * as i18n from '@/languages';
-import { networkSwitcherStore } from '@/state/networkSwitcher/networkSwitcher';
+
+import { networkSwitcherStore } from '../stores/networkSwitcher';
 
 export const SEARCH_BAR_HEIGHT = 48;
 const SEARCH_BAR_BORDER_RADIUS = SEARCH_BAR_HEIGHT / 2;
