+++ src/features/network/screens/NetworkSelector.tsx
@@ -34,7 +34,6 @@ import { AnimatedTextIcon } from '@/components/AnimatedComponents/AnimatedTextIc
 import { SPRING_CONFIGS, TIMING_CONFIGS } from '@/components/animations/animationConfigs';
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { GestureHandlerButton } from '@/components/buttons/GestureHandlerButton';
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { TapToDismiss } from '@/components/DappBrowser/control-panel/ControlPanel';
 import { EasingGradient } from '@/components/easing-gradient/EasingGradient';
 import {
@@ -57,18 +56,19 @@ import type Routes from '@/navigation/routesNames';
 import { type RootStackParamList } from '@/navigation/types';
 import { useBackendNetworksStore } from '@/state/backendNetworks/backendNetworks';
 import { ChainId } from '@/state/backendNetworks/types';
+import { THICK_BORDER_WIDTH, THICKER_BORDER_WIDTH } from '@/styles/constants';
+import deviceUtils, { DEVICE_WIDTH } from '@/utils/deviceUtils';
+import safeAreaInsetValues from '@/utils/safeAreaInsetValues';
+
+import { ChainImage } from '../components/ChainImage';
+import { SEARCH_BAR_HEIGHT, SearchBar } from '../components/SearchBar';
 import {
   customizeNetworksBannerStore,
   defaultPinnedNetworks,
   dismissCustomizeNetworksBanner,
   networkSwitcherStore,
   shouldShowCustomizeNetworksBanner,
-} from '@/state/networkSwitcher/networkSwitcher';
-import { THICK_BORDER_WIDTH, THICKER_BORDER_WIDTH } from '@/styles/constants';
-import deviceUtils, { DEVICE_WIDTH } from '@/utils/deviceUtils';
-import safeAreaInsetValues from '@/utils/safeAreaInsetValues';
-
-import { SEARCH_BAR_HEIGHT, SearchBar } from './components/SearchBar';
+} from '../stores/networkSwitcher';
 
 type RouteParams = RouteProp<RootStackParamList, 'NetworkSelector'>['params'];
