+++ src/features/network/stores/networkSwitcher.ts
@@ -1,9 +1,8 @@
 import { analytics } from '@/analytics';
 import { logger } from '@/logger';
 import { ChainId } from '@/state/backendNetworks/types';
+import { createRainbowStore } from '@/state/internal/createRainbowStore';
 import { useNonceStore } from '@/state/nonces';
-
-import { createRainbowStore } from '../internal/createRainbowStore';
 
 export const defaultPinnedNetworks = [ChainId.base, ChainId.mainnet, ChainId.optimism, ChainId.arbitrum, ChainId.polygon, ChainId.zora];
