+++ src/screens/mints/MintSheet.tsx
@@ -12,7 +12,6 @@ import { privateKeyToAccount } from 'viem/accounts';
 import { analytics } from '@/analytics';
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { HoldToAuthorizeButton } from '@/components/buttons';
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { ContactAvatar } from '@/components/contacts';
 import { useRainbowToastEnabled } from '@/components/rainbow-toast/useRainbowToastEnabled';
 import { Bleed, Box, ColorModeProvider, Column, Columns, Inline, Inset, Separator, Stack, Text } from '@/design-system';
@@ -22,6 +21,7 @@ import useENSAvatar from '@/features/ens/hooks/useENSAvatar';
 import { fetchReverseRecord } from '@/features/ens/utils/handlers';
 import GasSpeedButton from '@/features/gas/components/GasSpeedButton';
 import useGas from '@/features/gas/hooks/useGas';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { watchingAlert } from '@/features/wallet/utils/watchingAlert';
 import styled from '@/framework/ui/styled-thing';
 import { metadataPOSTClient } from '@/graphql';
