+++ src/screens/token-launcher/components/NetworkField.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { Bleed, Box, Inline, Text } from '@/design-system';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
