+++ src/screens/token-launcher/components/ReviewStep.tsx
@@ -3,11 +3,11 @@ import { ScrollView, StyleSheet } from 'react-native';
 
 import FastImage from 'react-native-fast-image';
 
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { isValidURLWorklet } from '@/components/DappBrowser/utils';
 import { Box, Text, TextShadow } from '@/design-system';
 import { type TextSize } from '@/design-system/components/Text/Text';
 import { convertAmountToNativeDisplay } from '@/features/currency/utils/nativeDisplay';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { abbreviateNumber, convertAmountToBalanceDisplay } from '@/helpers/utilities';
 import * as i18n from '@/languages';
 import { userAssetsStoreManager } from '@/state/assets/userAssetsStoreManager';
