+++ src/systems/funding/components/shared/ChainSelectorButton.tsx
@@ -1,8 +1,8 @@
 import React, { memo } from 'react';
 
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
-import { ChainImage } from '@/components/coin-icon/ChainImage';
 import { Box, Inline, Text, TextIcon } from '@/design-system';
+import { ChainImage } from '@/features/network/components/ChainImage';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { useBackendNetworksStore } from '@/state/backendNetworks/backendNetworks';
 import { type ChainId } from '@/state/backendNetworks/types';
